Update JQuery to `1.12.4`
So the `1.11.x` tree was deprecated (they were at `1.11.3` the code was using `1.11.1`) for multiple remote and XSS vulnerabilities. 

The `1.12.x` will be support into the future as things move to `3.0` and is mostly patching: http://blog.jquery.com/2016/01/08/jquery-2-2-and-1-12-released/

This *Shouldn't* break anything, but leaving as is opens a bunch of security holes.