Update libraries -current add integrity checking (SRI)
So this is almost identical to #7 but adds "[Subresource Integrity](https://www.srihash.org/)" checking to the `jquery` and `bootstrap` javascript. 

And can either be used in place of #7, or be rolled out after discussion.

`integrity=` are shasum's of known good versions of the `.js` files, thus preventing "MITM" and/or poisoning attacks. This is the current recommended production method for [Bootstrap](https://blog.getbootstrap.com/2016/07/25/bootstrap-3-3-7-released/#bootstrap-cdn) and [jquery](https://code.jquery.com/) for CDN deployed systems.
---
* JQuery
Bump `1.11.1` to `1.12.4`
So the `1.11.x` tree was deprecated (they were at `1.11.3` the code was using `1.11.1`) and had multiple remote and XSS vulnerabilities.
The `1.12.x` tree will be support into the future as things move to `3.0` and is mostly patching: http://blog.jquery.com/2016/01/08/jquery-2-2-and-1-12-released/
This Shouldn't break anything, but leaving as is leaves open a bunch of security holes.

* Bootstrap 3
Bump `3.3.6` to `3.3.7`
https://blog.getbootstrap.com/2016/07/25/bootstrap-3-3-7-released/
Also requires ` JQuery 1.12.4`

* html5shiv
Bump `3.7.2` to `3.7.3` for speed reasons (10x increase)
[Changelog](https://github.com/aFarkas/html5shiv/wiki):
```
v 3.3
complete refactoring by jdalton
huge performance improvement on createElement (more than 10 times faster compared to 3.2)
```

* Respond remains at `1.4.2`